+++ spring-blog/src/main/java/com/example/springblog/exceptions/SpringBlogException.java
@@ -1,0 +1,8 @@
+package com.example.springblog.exceptions;
+
+public class SpringBlogException extends RuntimeException {
+
+    public SpringBlogException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}